pyproject.toml
[project]
name = "pipecat-adk"
version = "0.1.0"
description = "Integrate Google ADK agents with Pipecat pipelines for real-time voice AI applications"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    { name = "Recruit41", email = "hello@recruit41.com" }
]
license = { text = "MIT" }

dependencies = [
    "pipecat-ai>=0.0.44",
    "google-adk>=0.1.0",
    "google-genai>=0.1.0",
    "google-cloud-texttospeech>=2.0.0",
    "google-cloud-speech>=2.0.0",
    "loguru>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
